cdk/cdk.out/asset.6909fffde436281a2f4822a22c37c9adb7fe995c7e9a4640ec906444f3507412/requirements-lambda.txt
# Core dependencies for Lambda
fastapi==0.104.1
mangum==0.17.0
pydantic==2.5.0

# AWS SDK
boto3==1.34.0

# Authentication
PyJWT==2.8.0
cryptography==41.0.7
requests==2.31.0
python-jose[cryptography]==3.3.0

# Data utilities
python-dateutil==2.8.2